requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pylint>=3.0.0
black>=23.0.0
pytest>=7.0.0